requirements.txt
streamlit~=1.51.0
langchain~=1.0.5
openai
python-dotenv~=1.2.1
duckduckgo-search
faiss-cpu
pypdf

langchain-community~=0.4.1
langchain-openai~=1.0.2
langchain-text-splitters~=1.0.0
langchain-core~=1.0.4